// src/types/company.ts
export interface Company {
  id: string;
  name: string;
  registreCommerce: string;
  country: string;
  city: string;
  location: string;
  sector: string;
  totalShares: number;
  availableShares: number;
  pricePerShare: number;
  previousPrice: number;
  logo: string;
  description: string;
  createdAt: string;
}